% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePic.R
\name{MakePic}
\alias{MakePic}
\title{Make dummy pictures}
\usage{
MakePic(n = 6, path = NULL, axes = FALSE, ...)
}
\arguments{
\item{n}{Number of pictures to plot}

\item{path}{Directory where the file wil be saved. If path=NULL, return a "magick-image" type.}

\item{axes}{Whether show the axes, same as axes in plot()}

\item{...}{more options, same as png/jpeg when path is NOT null. OR options in magick::image_graph(...) WHEN path=NULL}
}
\description{
Make dummy pictures
}
\examples{

# width, height and res pass to magick::image_graph()
figs=MakePic(axes = TRUE, res=72, width=400, height=400)
print(figs)

# save pictures in to path 'Figure'
# figs=MakePic(n=6, axes = TRUE, path ='Figure' )
# list.files('Figure')

}
