% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCanvas.R
\name{plotCanvas}
\alias{plotCanvas}
\title{Plot canvas for picture montage.}
\usage{
plotCanvas(
  width = 1200,
  height = 900,
  res = 96,
  col.bg = "white",
  TEST = FALSE,
  dev.off = TRUE
)
}
\arguments{
\item{width}{Width of the picture, see magick::image_graph()}

\item{height}{Height of the picture, see magick::image_graph()}

\item{res}{'res' of the picture, see magick::image_graph()}

\item{col.bg}{Color of the background}

\item{TEST}{Whether in test mode.}

\item{dev.off}{Whethere call dev.off(). If FALSE, you can add more options, such as title/drawing, into the canvas.}
}
\description{
The canvas is defined as rectangle, where xlim=c(0, 1) and ylim=c(0, 1).
Resolution, width and height are defined with function magick::image_graph().
}
\examples{

x=plotCanvas(TEST=TRUE, col.bg = 'pink')
print(x)

# --- plot canvas with title or background plot.
fig=plotCanvas(TEST=TRUE, col.bg = 'pink', dev.off = FALSE);
x=seq(0, 1, length.out=50); y = sin(x*12)/4 +0.25
lines(x,  y, col='red')
mtext(side=3, 'Sine plot', line=-5, cex=3, col='darkblue')
dev.off()
plot(fig)


}
